+++ packages/core/src/transform-block-directive.ts
@@ -1,15 +1,15 @@
 import type * as babel from '@babel/core';
 import * as t from '@babel/types';
-import { splitBlock } from './split';
-import type { DirectiveDefinition, StateContext } from './types';
+import { splitBlockDirective } from './split-block-directive';
+import type { BlockDirectiveDefinition, StateContext } from './types';
 
 function getValidDirectiveFromString(
   ctx: StateContext,
   string: string,
-): DirectiveDefinition | undefined {
-  for (let i = 0, len = ctx.options.directives.length; i < len; i++) {
-    const current = ctx.options.directives[i];
-    if (current.value === string) {
+): BlockDirectiveDefinition | undefined {
+  for (let i = 0, len = ctx.options.definitions.length; i < len; i++) {
+    const current = ctx.options.definitions[i];
+    if (current.type === 'block-directive' && current.directive === string) {
       return current;
     }
   }
@@ -19,7 +19,7 @@ function getValidDirectiveFromString(
 function getDefinitionFromDirectives(
   ctx: StateContext,
   path: babel.NodePath<t.BlockStatement>,
-): DirectiveDefinition | undefined {
+): BlockDirectiveDefinition | undefined {
   for (let i = 0, len = path.node.directives.length; i < len; i++) {
     const statement = path.node.directives[i].value.value;
     const directive = getValidDirectiveFromString(ctx, statement);
@@ -33,7 +33,7 @@ function getDefinitionFromDirectives(
 function getDefinitionFromFauxDirectives(
   ctx: StateContext,
   path: babel.NodePath<t.BlockStatement>,
-): DirectiveDefinition | undefined {
+): BlockDirectiveDefinition | undefined {
   for (let i = 0, len = path.node.body.length; i < len; i++) {
     const statement = path.node.body[i];
     if (
@@ -57,7 +57,7 @@ function getDefinitionFromFauxDirectives(
 function getDirectiveDefinitionFromBlock(
   ctx: StateContext,
   path: babel.NodePath<t.BlockStatement>,
-): DirectiveDefinition | undefined {
+): BlockDirectiveDefinition | undefined {
   const parent = path.getFunctionParent();
   if (parent && !parent.node.async) {
     return undefined;
@@ -70,12 +70,12 @@ function getDirectiveDefinitionFromBlock(
 
 function cleanBlockForDirectives(
   path: babel.NodePath<t.BlockStatement>,
-  definition: DirectiveDefinition,
+  definition: BlockDirectiveDefinition,
 ): void {
   const newDirectives: t.Directive[] = [];
   for (let i = 0, len = path.node.directives.length; i < len; i++) {
     const current = path.node.directives[i];
-    if (current.value.value !== definition.value) {
+    if (current.value.value !== definition.directive) {
       newDirectives.push(current);
     }
   }
@@ -84,7 +84,7 @@ function cleanBlockForDirectives(
 
 function cleanBlockForFauxDirectives(
   path: babel.NodePath<t.BlockStatement>,
-  definition: DirectiveDefinition,
+  definition: BlockDirectiveDefinition,
 ): void {
   const body = path.get('body');
   for (let i = 0, len = body.length; i < len; i++) {
@@ -93,7 +93,7 @@ function cleanBlockForFauxDirectives(
       t.isExpressionStatement(statement.node) &&
       t.isStringLiteral(statement.node.expression)
     ) {
-      if (statement.node.expression.value === definition.value) {
+      if (statement.node.expression.value === definition.directive) {
         statement.remove();
         return;
       }
@@ -101,15 +101,15 @@ function cleanBlockForFauxDirectives(
   }
 }
 
-export function transformBlock(
+export function transformBlockDirective(
   ctx: StateContext,
   path: babel.NodePath<t.BlockStatement>,
 ): void {
   const definition = getDirectiveDefinitionFromBlock(ctx, path);
-  if (!definition) {
+  if (!definition || definition.type === 'block-directive') {
     return;
   }
   cleanBlockForDirectives(path, definition);
   cleanBlockForFauxDirectives(path, definition);
-  splitBlock(ctx, path, definition);
+  splitBlockDirective(ctx, path, definition);
 }
