+++ packages/core/tests/block-directives/arrow-function.test.ts
