+++ packages/core/tests/block-directives/example.ts
@@ -2,9 +2,10 @@ import type * as compiler from '../../src';
 
 export const SERVER: compiler.Options = {
   key: 'my-example',
-  directives: [
+  definitions: [
     {
-      value: 'use server',
+      type: 'block-directive',
+      directive: 'use server',
       target: {
         source: 'my-example/server',
         kind: 'named',
@@ -12,15 +13,15 @@ export const SERVER: compiler.Options = {
       },
     },
   ],
-  functions: [],
   mode: 'server',
   env: 'development',
 };
 export const CLIENT: compiler.Options = {
   key: 'my-example',
-  directives: [
+  definitions: [
     {
-      value: 'use server',
+      type: 'block-directive',
+      directive: 'use server',
       target: {
         source: 'my-example/client',
         kind: 'named',
@@ -28,7 +29,6 @@ export const CLIENT: compiler.Options = {
       },
     },
   ],
-  functions: [],
   mode: 'client',
   env: 'development',
 };
