+++ packages/core/tests/block-directives/for-of-statement.test.ts
