+++ packages/core/tests/block-directives/function-declaration.test.ts
