+++ packages/core/tests/block-directives/function-expression.test.ts
