+++ packages/core/tests/block-directives/labeled-statement.test.ts
